Created task1.lex
Creates a flex specification to recognize tokens in the calculator language. Prints out an error message and exits if any unrecognized character is encountered in the input. Uses flex to compile your specification into an executable lexical analyzer that reads text from standard-in and prints the identified tokens to the screen, one token per line.