Created task3.lex
Enhance my flex specification so that in addition to printing out each encountered token, the lexical analyzer will also count the following:

The number of integers encountered
The number of operators encountered: +, -, *, /
The number of parentheses encountered: (, )
The number of equal signs encountered

The total counts should be printed to the screen after all input text has been tokenized. The counts do not need to be printed if an unrecognized character is encountered in the input (since the lexical analyzer should just terminate after issuing the error message).